+++ Paloma Mensajera/src/interfazGrafica/Principal.java
@@ -1,0 +1,12 @@
+package interfazGrafica;
+
+import javax.swing.JPanel;
+
+public class Principal extends JPanel {
+
+	public Principal() {
+		setLayout(null);
+
+	}
+
+}
